Improving argument parsing in find_orf.py.
Improving the CLI and argument parsing in the `main` function of
`find_orf.py`. No longer allowing a list of codons to be specified for
the start/stop codon options, because this can be confused with the
positional argument of a the sequence. This can be separated using `--`,
but it's cleaner to allow the start/stop codon flags to be used multiple
times and to append them to a list.